+++ http/src/routing/mod.rs
@@ -1,9 +1,5 @@
-mod path;
 mod route;
 mod route_display;
 
-pub use self::{
-    path::{Path, PathParseError, PathParseErrorType},
-    route::Route,
-    route_display::RouteDisplay,
-};
+pub use self::{route::Route, route_display::RouteDisplay};
+pub use twilight_http_ratelimiting::request::{Path, PathParseError, PathParseErrorType};