+++ model/src/guild/integration.rs
@@ -1,5 +1,6 @@
 use super::{IntegrationAccount, IntegrationApplication, IntegrationExpireBehavior};
 use crate::{
+    datetime::Timestamp,
     id::{GuildId, IntegrationId, RoleId},
     user::User,
 };
@@ -30,7 +31,7 @@ pub struct GuildIntegration {
     #[serde(skip_serializing_if = "Option::is_none")]
     pub subscriber_count: Option<u64>,
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub synced_at: Option<String>,
+    pub synced_at: Option<Timestamp>,
     #[serde(skip_serializing_if = "Option::is_none")]
     pub syncing: Option<bool>,
     #[serde(skip_serializing_if = "Option::is_none")]
@@ -43,12 +44,18 @@ mod tests {
         GuildIntegration, IntegrationAccount, IntegrationApplication, IntegrationExpireBehavior,
         IntegrationId, User,
     };
-    use crate::id::{ApplicationId, RoleId, UserId};
+    use crate::{
+        datetime::{Timestamp, TimestampParseError},
+        id::{ApplicationId, RoleId, UserId},
+    };
     use serde_test::Token;
+    use std::str::FromStr;
 
     #[allow(clippy::too_many_lines)]
     #[test]
-    fn test_guild_integration() {
+    fn test_guild_integration() -> Result<(), TimestampParseError> {
+        let synced_at = Timestamp::from_str("2021-01-01T01:01:01+00:00")?;
+
         let value = GuildIntegration {
             account: IntegrationAccount {
                 id: "abcd".to_owned(),
@@ -66,7 +73,7 @@ mod tests {
             revoked: Some(false),
             role_id: Some(RoleId::new(3).expect("non zero")),
             subscriber_count: Some(1337),
-            synced_at: Some("timestamp".to_owned()),
+            synced_at: Some(synced_at),
             syncing: Some(false),
             user: Some(User {
                 accent_color: None,
@@ -136,7 +143,7 @@ mod tests {
                 Token::U64(1337),
                 Token::Str("synced_at"),
                 Token::Some,
-                Token::Str("timestamp"),
+                Token::Str("2021-01-01T01:01:01.000000+00:00"),
                 Token::Str("syncing"),
                 Token::Some,
                 Token::Bool(false),
@@ -166,11 +173,15 @@ mod tests {
                 Token::StructEnd,
             ],
         );
+
+        Ok(())
     }
 
     #[allow(clippy::too_many_lines)]
     #[test]
-    fn test_guild_integration_complete() {
+    fn test_guild_integration_complete() -> Result<(), TimestampParseError> {
+        let synced_at = Timestamp::from_str("2021-01-01T01:01:01+00:00")?;
+
         let value = GuildIntegration {
             account: IntegrationAccount {
                 id: "abcd".to_owned(),
@@ -195,7 +206,7 @@ mod tests {
             revoked: Some(false),
             role_id: Some(RoleId::new(3).expect("non zero")),
             subscriber_count: Some(1337),
-            synced_at: Some("timestamp".to_owned()),
+            synced_at: Some(synced_at),
             syncing: Some(false),
             user: Some(User {
                 accent_color: None,
@@ -285,7 +296,7 @@ mod tests {
                 Token::U64(1337),
                 Token::Str("synced_at"),
                 Token::Some,
-                Token::Str("timestamp"),
+                Token::Str("2021-01-01T01:01:01.000000+00:00"),
                 Token::Str("syncing"),
                 Token::Some,
                 Token::Bool(false),
@@ -315,5 +326,7 @@ mod tests {
                 Token::StructEnd,
             ],
         );
+
+        Ok(())
     }
 }